Move parseParamsInPath to SkygearRequest
